pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "minsearch>=0.0.7",
    "openai>=2.15.0",
    "pydantic-ai>=1.43.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
